nix compile fixes Makefile additions
git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@287 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/ascent-world/SpellAuras.h
@@ -678,10 +678,10 @@ public:
 
 	bool HasMechanic(uint32 MechanicsType)
 	{
-		return m_spellProto->MechanicsType == MechanicsType ||
-			m_spellProto->EffectMechanic[0] == MechanicsType ||
-			m_spellProto->EffectMechanic[1] == MechanicsType ||
-			m_spellProto->EffectMechanic[2] == MechanicsType;
+		return (uint32)m_spellProto->MechanicsType == MechanicsType ||
+			(uint32)m_spellProto->EffectMechanic[0] == MechanicsType ||
+			(uint32)m_spellProto->EffectMechanic[1] == MechanicsType ||
+			(uint32)m_spellProto->EffectMechanic[2] == MechanicsType;
 	}
 	uint32 GetMechanic()
 	{
